fix: Convinced that nested virt needed for virtnet
Used to think there was a way to do this in MacOS natively. From what
I've read, only pre-catalina (2015) were there kernel extensions for
[tuntap](https://tuntaposx.sourceforge.net/).

In addition, it seems that WSL2 working with Layer2 protocols of PXE and
TFTP would required a kind of NAT setup that will become more difficult
to support than Vbox or Hyper-V.

Vbox obviously is something I want to avoid due to the predatory nature
of Oracle and how easy it is for some unknowing person to install the
proprietary extension pack.

Signed-off-by: Kevin Rineer <kevinrineer@proton.me>